requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
pymongo==4.1.1
python-dotenv==0.19.0